requirements.txt
streamlit>=1.45.0
openai>=1.77.0
python-dotenv>=1.1.0
faiss-cpu>=1.11.0
numpy>=1.23
beautifulsoup4>=4.13.4
tiktoken>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
